+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.1 - 2023-11-12
+
+### Fixed
+
+- Wait less time to restart the source by polling the tasks instead of waiting for them to be reactivated
+- Takes less time and memory to compute the timeout and streams to watch
+
 ## 2.0.0 - 2023-11-05
 
 ### Added